Made some changes after today's class.
I would really appreaciate it if you could consider both of my versions. Since I've spent a lot of time and effort on this course and lab2 preparation. Thank you very much for your time and consideration! (I also uploaded one preparation .ipynb file which shows my effort on preparing for this lab and work on this course.